feat: input을 uncontrolled로 변경
