+++ src/components/index.ts
@@ -1,0 +1,1 @@
+export { SearchPage } from "./SearchPage";
